backend/requirements.txt
Django==5.0.9
Pillow==10.3.0
cryptography==42.0.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.1.19
django-filter==24.2
django-cors-headers==4.3.1
drf-spectacular==0.27.2
drf-flex-fields==1.0.2
python-dotenv==1.0.1
django-storages[s3]==1.14.2
google-auth==2.29.0
sentry-sdk[django]==2.20.0
argon2-cffi==23.1.0
redis==5.0.8
